Use BN for balances and rename constants
