+++ src/components/DepositWidget/Row.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
+import BN from 'bn.js'
 
 import { TokenBalanceDetails } from 'types'
 import unknownTokenImg from 'img/unknown-token.png'
@@ -20,6 +21,10 @@ export interface RowProps {
   tokenBalances: TokenBalanceDetails
 }
 
+function formatBN(number: BN): string {
+  return number.toString()
+}
+
 export const Row: React.FC<RowProps> = (props: RowProps) => {
   const tokenBalances = props.tokenBalances
 
@@ -29,16 +34,17 @@ export const Row: React.FC<RowProps> = (props: RowProps) => {
         <img src={tokenBalances.image} alt={tokenBalances.name} onError={loadFallbackTokenImage} />
       </td>
       <td>{tokenBalances.name}</td>
-      <td>{tokenBalances.exchangeWallet}</td>
-      <td>{tokenBalances.pendingDeposits}</td>
-      <td>{tokenBalances.pendingWithdraws}</td>
+      <td>{formatBN(tokenBalances.exchangeWallet)}</td>
+      <td>{formatBN(tokenBalances.pendingDeposits)}</td>
+      <td>{formatBN(tokenBalances.pendingWithdraws)}</td>
       <td>
-        {!tokenBalances.enabled && <button className="success">✓ Enable {tokenBalances.symbol}</button>}
-        {tokenBalances.enabled && (
+        {tokenBalances.enabled ? (
           <>
             <button>+ Deposit</button>
             <button className="danger">- Withdraw</button>
           </>
+        ) : (
+          <button className="success">✓ Enable {tokenBalances.symbol}</button>
         )}
       </td>
     </WrapperRow>
